rag-gemini-pdf/requirements.txt
python-dotenv
qdrant-client
openai
google-generativeai
PyMuPDF
python-pptx
python-docx
tiktoken
streamlit
